backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dateutil>=2.8
pandas>=2.0
